Update the MWT library classes to accept int/double instead of short/float, to reduce explicit casting elsewhere
'ref' parameters were also changed, but returns and 'out' parameters were left as-is since an implicit cast is possible

+++ MolecularWeightCalculatorLib/COMInterfaces/ICompound.cs
@@ -11,11 +11,11 @@ namespace MolecularWeightCalculator.COMInterfaces
         // Ignore Spelling: interop
 
         string ConvertToEmpirical();
-        bool ElementPresent(short atomicNumber);
+        bool ElementPresent(int atomicNumber);
         string ExpandAbbreviations();
-        double GetAtomCountForElement(short atomicNumber);
-        double GetPercentCompositionForElement(short atomicNumber);
-        string GetPercentCompositionForElementAsString(short atomicNumber, bool includeStandardDeviation = true);
+        double GetAtomCountForElement(int atomicNumber);
+        double GetPercentCompositionForElement(int atomicNumber);
+        string GetPercentCompositionForElementAsString(int atomicNumber, bool includeStandardDeviation = true);
 
         /// <summary>
         /// Get the percent composition for all elements in an empirical formula. This implementation is specifically for COM interop support
